+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/test/GeoPackageExample.java
@@ -9,7 +9,7 @@ import android.content.pm.PackageManager.NameNotFoundException;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Environment;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 
 import junit.framework.TestCase;