+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/features/custom/NumberFeaturesTile.java
@@ -6,7 +6,7 @@ import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.Rect;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.util.TypedValue;
 
 import mil.nga.geopackage.GeoPackageException;